Move custodial wallet source code to subfolder
