--- .sample.env
@@ -1,7 +1,0 @@
-DB_HOST=localhost
-DB_USERNAME=cosmos-wallet
-DB_PASSWORD=cosmos-wallet
-DB_NAME=cosmos-wallet
-DB_PORT=5432
-
-APP_ENVIRONMENT=DEV # DEV, PROD
