+++ custodial/README.md
@@ -1,0 +1,11 @@
+# cosmos-wallet
+
+Wallet modules for Cosmos Network - AssetMantle
+
+# Production deployment
+
+Set env `APP_ENVIRONMENT` to `PROD`
+
+# Design decisions
+
+- `o` prefix is used where there might be conflict, for e.g. package names like errorso, logo
