+++ custodial/app/stage/appinit/init.go
@@ -1,0 +1,13 @@
+package appinit
+
+import (
+	"github.com/MyriadFlow/cosmos-wallet/custodial/app/stage/appinit/dbconinit"
+	"github.com/MyriadFlow/cosmos-wallet/custodial/app/stage/appinit/dbmigrate"
+	"github.com/MyriadFlow/cosmos-wallet/custodial/app/stage/appinit/logoinit"
+)
+
+func Init() {
+	logoinit.Init()
+	dbconinit.Init()
+	dbmigrate.Migrate()
+}
