+++ custodial/models/user/hooks.go
@@ -6,13 +6,13 @@ import (
 	"gorm.io/gorm"
 )
 
-func (u *User) BeforeSave(tx *gorm.DB) (err error) {
+func (u *CustodialUser) BeforeSave(tx *gorm.DB) (err error) {
 	hexMnemonic := "0x" + hex.EncodeToString([]byte(u.Mnemonic))
 	u.Mnemonic = hexMnemonic
 	return nil
 }
 
-func (u *User) AfterFind(tx *gorm.DB) (err error) {
+func (u *CustodialUser) AfterFind(tx *gorm.DB) (err error) {
 	plainMnemonic, err := hex.DecodeString(string(u.Mnemonic[2:]))
 	if err != nil {
 		return err
